refactor(sandbox-fc): promote shared paths to global flags and expose exec options
Move firecracker/kernel/rootfs from positional args duplicated across
subcommands to global --flags. Convert base_dir to --base-dir flag.
Expose vcpu-count, memory-mb, and timeout as configurable exec options
instead of hardcoded values.

+++ crates/sandbox/src/config.rs
@@ -1,6 +1,6 @@
 pub struct ResourceLimits {
     pub cpu_count: u32,
-    pub memory_mb: u64,
+    pub memory_mb: u32,
     pub timeout_secs: u64,
 }
